test(product/repository): Create and GetAll cases

+++ internal/products/domain/product_domain.go
@@ -7,16 +7,16 @@ import (
 type Product struct {
 	Id                             int64   `json:"id"`
 	Description                    string  `json:"description"`
-	ExpirationRate                 int     `json:"expiration_rate"`
-	FreezingRate                   int     `json:"freezing_rate"`
+	ExpirationRate                 int64     `json:"expiration_rate"`
+	FreezingRate                   int64     `json:"freezing_rate"`
 	Height                         float64 `json:"height"`
 	Length                         float64 `json:"length"`
 	NetWeight                      float64 `json:"netweight"`
 	ProductCode                    string  `json:"product_code"`
 	RecommendedFreezingTemperature float64 `json:"recommended_freezing_temperature"`
 	Width                          float64 `json:"width"`
-	ProductTypeId                  int     `json:"product_type_id"`
-	SellerId                       int     `json:"seller_id"`
+	ProductTypeId                  int64     `json:"product_type_id"`
+	SellerId                       int64     `json:"seller_id"`
 }
 
 type Repository interface {
@@ -37,16 +37,16 @@ type Service interface {
 
 type RequestProducts struct {
 	Description                    string  `json:"description" binding:"required"`
-	ExpirationRate                 int     `json:"expiration_rate" binding:"required"`
-	FreezingRate                   int     `json:"freezing_rate" binding:"required"`
+	ExpirationRate                 int64     `json:"expiration_rate" binding:"required"`
+	FreezingRate                   int64     `json:"freezing_rate" binding:"required"`
 	Height                         float64 `json:"height" binding:"required"`
 	Length                         float64 `json:"length" binding:"required"`
 	NetWeight                      float64 `json:"netweight" binding:"required"`
 	ProductCode                    string  `json:"product_code" binding:"required"`
 	RecommendedFreezingTemperature float64 `json:"recommended_freezing_temperature" binding:"required"`
 	Width                          float64 `json:"width" binding:"required"`
-	ProductTypeId                  int     `json:"product_type_id" binding:"required"`
-	SellerId                       int     `json:"seller_id" binding:"required"`
+	ProductTypeId                  int64     `json:"product_type_id" binding:"required"`
+	SellerId                       int64     `json:"seller_id" binding:"required"`
 }
 
 type RequestProductId struct {
@@ -55,14 +55,14 @@ type RequestProductId struct {
 
 type RequestProductsUpdated struct {
 	Description                    string  `json:"description"`
-	ExpirationRate                 int     `json:"expiration_rate"`
-	FreezingRate                   int     `json:"freezing_rate"`
+	ExpirationRate                 int64     `json:"expiration_rate"`
+	FreezingRate                   int64     `json:"freezing_rate"`
 	Height                         float64 `json:"height"`
 	Length                         float64 `json:"length"`
 	NetWeight                      float64 `json:"netweight"`
 	ProductCode                    string  `json:"product_code"`
 	RecommendedFreezingTemperature float64 `json:"recommended_freezing_temperature"`
 	Width                          float64 `json:"width"`
-	ProductTypeId                  int     `json:"product_type_id"`
-	SellerId                       int     `json:"seller_id"`
+	ProductTypeId                  int64     `json:"product_type_id"`
+	SellerId                       int64     `json:"seller_id"`
 }
